Refactoring class of model and frontend
