+++ luiz/to-do-list/app/index.html
@@ -35,6 +35,9 @@
   <script src="https://cdnjs.cloudflare.com/ajax/libs/angular-ui-router/0.2.8/angular-ui-router.min.js"></script>
 
   <script src="/js/app.js"></script>
+  <script src="/js/service/todoService.js"></script>
+  <script src="/js/controllers/tarefasController.js"></script>
+  <script src="/js/controllers/listaTarefasCtrl.js"></script>
   
 </body>
 </html>
