+++ luiz/to-do-list/util/util.py
@@ -1,0 +1,21 @@
+import json
+
+
+def date_handler(obj):
+  if hasattr(obj, 'isoformat'):
+    return obj.isoformat()
+  elif hasattr(obj, 'email'):
+    return obj.email()
+
+  return obj
+
+def data2json(data):
+  return json.dumps(
+    data,
+    default=date_handler,
+    indent=2,
+    separators=(',', ': '),
+    ensure_ascii=False)
+
+def json2data(jsonStr):
+  return json.loads(jsonStr)
